requirements.txt
# 🎧 Audiobook Downloader Dependencies
# Основные зависимости для работы загрузчика аудиокниг

# YouTube и видео загрузка
yt-dlp>=2023.12.30

# Поиск в Google (опционально - для улучшенной версии)
# Может не работать в некоторых версиях, используйте простую версию
googlesearch-python>=1.2.3

# HTTP запросы
requests>=2.31.0

# Работа с путями (для совместимости)
pathlib2>=2.3.7

# Красивый интерфейс в терминале
rich>=13.5.0

# Дополнительные утилиты
python-dateutil>=2.8.2
colorama>=0.4.6
